Client side (JS) translation
Ugggllyyyyyy and hackkyyy way of doing it. But i'm tiiiired and
don't want to rewrite huge parts of the App to make it works
in a cleaner way ! (aka promises, js classes, ...)

I first added a service to return the translated json, to get it
later through an ajax call... but then I saw the mess in loading the
scriptasaurus and editor files. The whole thing was loaded before the
ajax call could return, and the translation was empty.
And the way the code is made, I couldn't make a promise to wait...

Sooooooo : I gave up and used the ugly way : a json loaded directly
through PHP inside a <script> tag, yay !